+++ cli/cli/kurtosis_gateway/server/api_container_gateway/api_container_gateway_service_server.go
@@ -84,18 +84,9 @@ func (service *ApiContainerGatewayServiceServer) ExecuteModule(ctx context.Conte
 	return remoteApiContainerResponse, nil
 }
 
-func (service *ApiContainerGatewayServiceServer) ExecuteStartosisScript(ctx context.Context, args *kurtosis_core_rpc_api_bindings.ExecuteStartosisScriptArgs) (*kurtosis_core_rpc_api_bindings.ExecuteStartosisResponse, error) {
-	remoteApiContainerResponse, err := service.remoteApiContainerClient.ExecuteStartosisScript(ctx, args)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, errorCallingRemoteApiContainerFromGateway)
-	}
-
-	return remoteApiContainerResponse, nil
-}
-
-func (service *ApiContainerGatewayServiceServer) ExecuteKurtosisScript(args *kurtosis_core_rpc_api_bindings.ExecuteStartosisScriptArgs, streamToWriteTo kurtosis_core_rpc_api_bindings.ApiContainerService_ExecuteKurtosisScriptServer) error {
-	logrus.Debug("Executing Kurtosis script")
-	streamToReadFrom, err := service.remoteApiContainerClient.ExecuteKurtosisScript(streamToWriteTo.Context(), args)
+func (service *ApiContainerGatewayServiceServer) RunStarlarkScript(args *kurtosis_core_rpc_api_bindings.RunStarlarkScriptArgs, streamToWriteTo kurtosis_core_rpc_api_bindings.ApiContainerService_RunStarlarkScriptServer) error {
+	logrus.Debug("Executing Starlark script")
+	streamToReadFrom, err := service.remoteApiContainerClient.RunStarlarkScript(streamToWriteTo.Context(), args)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred starting the execution of Kurtosis code")
 	}
@@ -105,23 +96,14 @@ func (service *ApiContainerGatewayServiceServer) ExecuteKurtosisScript(args *kur
 	return nil
 }
 
-func (service *ApiContainerGatewayServiceServer) ExecuteStartosisModule(ctx context.Context, args *kurtosis_core_rpc_api_bindings.ExecuteStartosisModuleArgs) (*kurtosis_core_rpc_api_bindings.ExecuteStartosisResponse, error) {
-	remoteApiContainerResponse, err := service.remoteApiContainerClient.ExecuteStartosisModule(ctx, args)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, errorCallingRemoteApiContainerFromGateway)
-	}
-
-	return remoteApiContainerResponse, nil
-}
-
-func (service *ApiContainerGatewayServiceServer) ExecuteKurtosisModule(args *kurtosis_core_rpc_api_bindings.ExecuteStartosisModuleArgs, streamToWriteTo kurtosis_core_rpc_api_bindings.ApiContainerService_ExecuteKurtosisModuleServer) error {
-	logrus.Debugf("Executing Kurtosis module '%s'", args.GetModuleId())
-	streamToReadFrom, err := service.remoteApiContainerClient.ExecuteKurtosisModule(streamToWriteTo.Context(), args)
+func (service *ApiContainerGatewayServiceServer) RunStarlarkPackage(args *kurtosis_core_rpc_api_bindings.RunStarlarkPackageArgs, streamToWriteTo kurtosis_core_rpc_api_bindings.ApiContainerService_RunStarlarkPackageServer) error {
+	logrus.Debugf("Executing Starlark package '%s'", args.GetPackageId())
+	streamToReadFrom, err := service.remoteApiContainerClient.RunStarlarkPackage(streamToWriteTo.Context(), args)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred starting the execution of Kurtosis code")
 	}
 	if err := service.forwardKurtosisExecutionStream(streamToReadFrom, streamToWriteTo); err != nil {
-		return stacktrace.Propagate(err, "Error forwarding stream from Kurtosis core back to the user while executing module '%s'", args.GetModuleId())
+		return stacktrace.Propagate(err, "Error forwarding stream from Kurtosis core back to the user while executing package '%s'", args.GetPackageId())
 	}
 	return nil
 }
@@ -446,9 +428,9 @@ func (service *ApiContainerGatewayServiceServer) idempotentKillRunningConnection
 
 func (service *ApiContainerGatewayServiceServer) forwardKurtosisExecutionStream(streamToReadFrom grpc.ClientStream, streamToWriteTo grpc.ServerStream) error {
 	for {
-		kurtosisExecutionResponseLine := new(kurtosis_core_rpc_api_bindings.KurtosisExecutionResponseLine)
+		starlarkRunResponseLine := new(kurtosis_core_rpc_api_bindings.StarlarkRunResponseLine)
 		// RecvMsg blocks until either a message is received or an error is thrown
-		readErr := streamToReadFrom.RecvMsg(kurtosisExecutionResponseLine)
+		readErr := streamToReadFrom.RecvMsg(starlarkRunResponseLine)
 		if readErr == io.EOF {
 			logrus.Debug("Finished reading from the Kurtosis response line stream.")
 			return nil
@@ -457,7 +439,7 @@ func (service *ApiContainerGatewayServiceServer) forwardKurtosisExecutionStream(
 			return stacktrace.Propagate(readErr, "Error reading Kurtosis execution lines from Kurtosis core stream")
 		}
 
-		if writeErr := streamToWriteTo.SendMsg(kurtosisExecutionResponseLine); writeErr != nil {
+		if writeErr := streamToWriteTo.SendMsg(starlarkRunResponseLine); writeErr != nil {
 			return stacktrace.Propagate(readErr, "Received a Kurtosis execution line but failed forwarding it back to the user")
 		}
 	}