+++ core/server/commons/enclave_data_directory/enclave_data_directory.go
@@ -6,7 +6,7 @@
 package enclave_data_directory
 
 import (
-	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_modules/git_module_content_provider"
+	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_packages/git_package_content_provider"
 	"github.com/kurtosis-tech/stacktrace"
 	"path"
 )
@@ -16,13 +16,13 @@ const (
 	// This will replace artifactCacheDirname
 	artifactStoreDirname = "artifact-store"
 
-	// The name of the directory INSIDE THE ENCLAVE DATA DIR where startosis modules will be stored
-	startosisModuleStoreDirname = "startosis-modules"
+	// The name of the directory INSIDE THE ENCLAVE DATA DIR where Starlark packages will be stored
+	startosisPackageStoreDirname = "startosis-packages"
 
-	// The name of the directory INSIDE THE ENCLAVE DATA DIR where temporary modules will be stored
+	// The name of the directory INSIDE THE ENCLAVE DATA DIR where temporary packages will be stored
 	// We place the temp folder here so that the move to the final destination is atomic
 	// Move from places outside of the enclave data dir are not atomic as they're over the network
-	tmpModuleStoreDirname = "tmp-startosis-modules"
+	tmpPackageStoreDirname = "tmp-startosis-packages"
 )
 
 // A directory containing all the data associated with a certain enclave (i.e. a Docker subnetwork where services are spun up)
@@ -45,16 +45,16 @@ func (dir EnclaveDataDirectory) GetFilesArtifactStore() (*FilesArtifactStore, er
 	return newFilesArtifactStore(absoluteDirpath, relativeDirpath), nil
 }
 
-func (dir EnclaveDataDirectory) GetGitModuleContentProvider() (*git_module_content_provider.GitModuleContentProvider, error) {
-	moduleStoreDirpath := path.Join(dir.absMountDirpath, startosisModuleStoreDirname)
-	if err := ensureDirpathExists(moduleStoreDirpath); err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred ensuring the Startosis module store dirpath '%v' exists.", moduleStoreDirpath)
+func (dir EnclaveDataDirectory) GetGitPackageContentProvider() (*git_package_content_provider.GitPackageContentProvider, error) {
+	packageStoreDirpath := path.Join(dir.absMountDirpath, startosisPackageStoreDirname)
+	if err := ensureDirpathExists(packageStoreDirpath); err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred ensuring the Starlark package store dirpath '%v' exists.", packageStoreDirpath)
 	}
 
-	tempModuleStoreDirpath := path.Join(dir.absMountDirpath, tmpModuleStoreDirname)
-	if err := ensureDirpathExists(tempModuleStoreDirpath); err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred ensuring the Startosis temporary module store dirpath '%v' exists.", tempModuleStoreDirpath)
+	tempPackageStoreDirpath := path.Join(dir.absMountDirpath, tmpPackageStoreDirname)
+	if err := ensureDirpathExists(tempPackageStoreDirpath); err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred ensuring the Starlark temporary package store dirpath '%v' exists.", tempPackageStoreDirpath)
 	}
 
-	return git_module_content_provider.NewGitModuleContentProvider(moduleStoreDirpath, tempModuleStoreDirpath), nil
+	return git_package_content_provider.NewGitPackageContentProvider(packageStoreDirpath, tempPackageStoreDirpath), nil
 }